refactor(cli): split commands into modules behind a Command trait
main.rs had grown to 1862 lines holding every command's options, pipeline,
rendering, and tests. Split it into one module per namespace and give each
command an explicit contract.

The Command trait runs in three steps: run, render, verdict. Keeping verdict
after render makes an invariant that was previously implicit in statement
order enforceable by the type — --fail-on and --fail-if-below decide the exit
code but can never suppress the report explaining the decision.

Context carries the working directory and config loading, replacing five
scattered std::env::current_dir().expect calls. Built-ins keep owning config
discovery, since catalog merge runs as a Git merge driver and version answers
before any project exists.

The shape deliberately mirrors a binary plugin command (ADR 018) so a command
that ever moves behind that protocol would swap its context and renderer
rather than its body. Built-ins stay compiled in: reaching them without
loading configuration or resolving plugins is what guarantees a broken plugin
cannot affect them (ADR 017).

Behavior is unchanged. Verified by diffing 19 invocations — every command,
both JSON modes, the error paths, and the benchmark timing marker — against a
binary built from the pre-split source: identical output and exit codes.

catalog convert now returns a report rendered in render instead of printing
mid-loop; the emitted text is unchanged.
